backend/requirements.txt
Django==6.0.2
django-allauth==65.14.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.48.0
pillow==12.1.1
requests==2.32.5
